Title: Landing page templating done
Description:
Completed styling for landing page, probably messed up the checklist page, but will get to that later.
Next is just to complete the regular note page for the sake of completing the coding challenge.

Moved the header into the todos/index template. I'll have to recreate the checklist functionality
later since the "regular note" piece is next.

Added right side checkbox and left side check mark identifier for "checklist" on the list of
todos such that the user can pick whether they want a regular note or a checklist.

Moved the footer into todos/checklist because it doesn't really belong in the todos/index.

Hella css changes. Removed clickability of the individual todos on the left side, added
css for pretty much everything on the todos/index, it will probably mess up the checklist
level css but oh well, i'll get to that later.