Title: complete edit regular note todo functionlity
Description:
Added functionality for editing regular notes.

Added styles to make sure the look stays the same when adding
a new one and editing an old one.

Added async: true to note relationship in the Todo model.

Got rid of useless note route. Added "todo" resource that uses
dynamic segements to map to the TodoRoute with the todo_id

Added "onInitialize" for autosize options so it would render the
correct size when editing a todo note.

Disabled the check boxes that allow creating a checklist. Won't have
that functionality until after I turn this project in.

changed label to link-to for editing a todo.

added a new template for editing a todo. it will also be the template
for editing the checklist todos.

Laid a little bit of groundwork for future work on checklists.
changed function from editTodo to editNote.

Added function saveTodo to save a todo after editing.